pyproject.toml
[tool.poetry]
name = "ai-debate"
version = "1.0.0"
description = "AI-driven multi-agent debate system for structured argumentation and analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ai_debate"}]

[tool.poetry.dependencies]
python = "^3.11"
google-genai = "^0.8.0"
google-cloud-storage = "^2.10.0"
google-cloud-aiplatform = "^1.35.0"
aiohttp = "^3.8.0"
asyncio = "^3.4.3"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true